Adjusted a makefile

+++ cpp02/ex03/Point.hpp
@@ -6,7 +6,7 @@
 /*   By: ouel-bou <ouel-bou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/27 13:06:10 by ouel-bou          #+#    #+#             */
-/*   Updated: 2025/04/27 14:47:57 by ouel-bou         ###   ########.fr       */
+/*   Updated: 2025/05/15 12:54:10 by ouel-bou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,6 @@
 
 class	Point
 {
-	
 	Fixed const	x;
 	Fixed const	y;
 
@@ -26,10 +25,8 @@ public:
 	Point(const Point& original);
 	Point&	operator=(const Point& original);
 	~Point(void);
-	int	getXVal(void) const;
-	int	getYVal(void) const;
-
-
+	float	getXVal(void) const;
+	float	getYVal(void) const;
 };
 
 bool	bsp(Point const a, Point const b, Point const c, Point const point);
